REFACTOR tests into 'Unit' & 'Feature' namespaces
ADD use of orchestra/testbench.
