+++ composer.json
@@ -22,6 +22,7 @@
     },
     "require-dev": {
         "phpunit/phpunit": ">=5.7.0",
+        "orchestra/testbench": ">=3.8",
         "scrutinizer/ocular": "^1.8"
     },
     "autoload": {
