+++ tests/TestCase.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Sfneal\Helpers\Arrays\Tests;
+
+class TestCase extends \Orchestra\Testbench\TestCase
+{
+
+}
